Reuse ExtraHddBodyEffectorLiveTest
